requirements.txt
langchain==1.2.4
langchain-openai==1.1.7
langchain-community==0.4.1
python-dotenv==1.2.1
rich==14.2.0
langchain-ollama==1.0.1
jsonschema==4.26.0
prompt_toolkit==3.0.52
textual==7.4.0
textual-autocomplete==4.0.6
psycopg[binary]==3.2.13
pgvector==0.4.2
unstructured==0.18.31
pypdf==6.6.2
python-docx==1.2.0
python-pptx==1.0.2
odfpy==1.4.1
striprtf==0.0.29
extract-msg==0.55.0
beautifulsoup4==4.13.5
PyYAML==6.0.3
